% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencltools_ldflags.R
\name{opencltoolsLdFlags}
\alias{opencltoolsLdFlags}
\title{Linker flags for downstream \code{LinkingTo: opencltools} packages}
\usage{
opencltoolsLdFlags()
}
\value{
A character scalar of linker flags (\code{-L... -lopencltools}).
}
\description{
Returns a \code{PKG_LIBS} fragment so a downstream package can resolve
\code{openclPort::} symbols from this package's shared library instead of
duplicating C++ sources such as \file{kernel_loader.cpp}.
}
\details{
Use together with \code{\link{opencltools}} headers
(\code{#include <opencltools/openclPort.h>}) and your own \file{configure}
logic for OpenCL SDK \code{-I} paths when \code{USE_OPENCL} is defined.

Typical \file{src/Makevars} usage (Unix):
\preformatted{
OPENCLTOOLS_LIBS = $(shell $(R_HOME)/bin/Rscript -e "opencltools::opencltoolsLdFlags()")
PKG_LIBS = $(OPENCLTOOLS_LIBS) $(PKG_LIBS)
}

On Windows, add the same \code{OPENCLTOOLS_LIBS} line to \file{src/Makevars.win}
(or merge into the \file{Makevars} block written by \file{configure.win}).
}
\examples{
############################ Start of opencltoolsLdFlags example ########################

flags <- opencltoolsLdFlags()
flags
grepl("-lopencltools", flags)

###############################################################################
## End of opencltoolsLdFlags example
###############################################################################
}
