% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_kernel_dependency_tags.R
\name{print.opencl_dependency_tags}
\alias{print.opencl_dependency_tags}
\title{Print Dependency Tag Attachment Results}
\usage{
\method{print}{opencl_dependency_tags}(x, max_rows = 50, ...)
}
\arguments{
\item{x}{Result of \link{attach_kernel_dependency_tags}.}

\item{max_rows}{Maximum printed rows per table section.}

\item{...}{Forwarded to \code{print.data.frame}.}
}
\value{
Called for side effects. Prints tag tables (and optional
cycle/unresolved reports) from \code{\link{attach_kernel_dependency_tags}}.
Invisibly returns \code{x}, a \code{list} with S3 class
\code{"opencl_dependency_tags"} (same structure as
\code{attach_kernel_dependency_tags()}: \code{ok}, \code{message},
\code{sorted}, \code{unresolved}, \code{cycles}, \code{tags}, \ldots).
}
\description{
Print output from \link{attach_kernel_dependency_tags}.
}
\examples{
############################ Start of attach_kernel_dependency_tags example ########################

lib_dir <- system.file("cl/nmath_small", package = "opencltools")
res <- attach_kernel_dependency_tags(lib_dir, dry_run = TRUE)
res$ok
print(res)

###############################################################################
## End of attach_kernel_dependency_tags example
###############################################################################
}
